server/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
django-extensions
django-cors-headers==3.7.0
django-admin-sortable2==0.7.8
django-filter==2.4.0
requests
espn_api
spacy==2.3.5
giphy_client
wolframalpha
praw==7.2.0
